+++ soybean.config.js
@@ -1,12 +1,13 @@
 
-import { Soybean, handlers } from './build/lib/exports.js'
+import { Soybean } from './build/exports/index.js'
+import handlers from './build/exports/handlers.js'
 
 export default Soybean({
     cp: {
         http: {
             command: ['http-server'],
             cwd: './',
-            stdout: 'all'
+            stdout: 'none'
         }
     },
     terminal: {
@@ -15,7 +16,14 @@ export default Soybean({
         handlers: {
             'test': handlers.handle((e) => {
                 console.log(`test command, argv:`, e.terminal.argvRaw || 'empty')
-            })
+            }),
+            'npm-init': handlers.group([
+                handlers.shell.spawn(['npm', 'init'], {
+                    stdio: 'takeover',
+                    cwd: './test',
+                    shell: 'cmd.exe'
+                })
+            ])
         }
     }
 })