+++ src/lib/events/handlers.ts
@@ -1,0 +1,3 @@
+
+export * as misc from './handlers/misc.js'
+export * as shell from './handlers/shell.js'